// pro/src/apiClient/adage/models/EducationalInstitutionWithBudgetResponseModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type EducationalInstitutionWithBudgetResponseModel = {
  budget: number;
  city: string;
  id: number;
  institutionType: string;
  name: string;
  phoneNumber: string;
  postalCode: string;
};